Generate methods to reset params (#84)
After merging PR #80, optional params always needs to be set.
In case of the optional parameter of type UUID, the only option is to set the value to an empty string. This results in a request to ACS with the optional parameter in the request with an empty string. ACS throws an error because it can't find the associated id in the database.

Actual change is only in generate.go. Rest of the changes are in generated code.